requirements.txt
pywebview[qt]
